Move faq logic out of main tino pkg/faq
